+++ src/main/java/io/scalecube/socketio/session/SessionStorage.java
@@ -10,25 +10,19 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.scalecube.socketio.storage;
+package io.scalecube.socketio.session;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.SocketAddress;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 import io.netty.channel.Channel;
 import io.scalecube.socketio.TransportType;
 import io.scalecube.socketio.packets.ConnectPacket;
 import io.scalecube.socketio.pipeline.UnsupportedTransportTypeException;
-import io.scalecube.socketio.session.FlashSocketSession;
-import io.scalecube.socketio.session.ManagedSession;
-import io.scalecube.socketio.session.SessionDisconnectHandler;
-import io.scalecube.socketio.session.JsonpPollingSession;
-import io.scalecube.socketio.session.WebSocketSession;
-import io.scalecube.socketio.session.XHRPollingSession;
-import io.scalecube.socketio.storage.memoizer.Computable;
-import io.scalecube.socketio.storage.memoizer.MemoizerConcurrentMap;
 
 /**
  *
@@ -39,7 +33,7 @@ public class SessionStorage {
 
   private final Logger log = LoggerFactory.getLogger(getClass());
 
-  private final MemoizerConcurrentMap<String, ManagedSession> sessionsMemoizer = new MemoizerConcurrentMap<>();
+  private final ConcurrentMap<String, ManagedSession> sessions = new ConcurrentHashMap<>();
 
   private final int localPort;
 
@@ -48,18 +42,18 @@ public class SessionStorage {
   }
 
   public boolean containSession(final String sessionId) {
-    return sessionsMemoizer.containsKey(sessionId);
+    return sessions.containsKey(sessionId);
   }
 
   public void removeSession(final String sessionId) {
-    sessionsMemoizer.remove(sessionId);
+    sessions.remove(sessionId);
   }
 
   public ManagedSession getSession(final ConnectPacket connectPacket,
                                     final Channel channel,
                                     final SessionDisconnectHandler disconnectHandler) throws Exception {
 
-    ManagedSession session = createSession(connectPacket, channel, disconnectHandler, null);
+    ManagedSession session = getOrCreateSession(connectPacket, channel, disconnectHandler, null);
 
     // If transport protocol was changed then remove old session and create new one instead
     if (connectPacket.getTransportType() != session.getTransportType()) {
@@ -70,13 +64,29 @@ public class SessionStorage {
 
       final String sessionId = connectPacket.getSessionId();
       removeSession(sessionId);
-      session = createSession(connectPacket, channel, disconnectHandler, session.getTransportType());
+      session = getOrCreateSession(connectPacket, channel, disconnectHandler, session.getTransportType());
 
       if (log.isDebugEnabled())
         log.debug("{} transport type {} session was upgraded to new transport type {} and session {}",
             oldTransportType.name(), oldSessionId, session.getTransportType().name(), session.getSessionId());
     }
 
+    return session;
+  }
+
+  private ManagedSession getOrCreateSession(final ConnectPacket connectPacket,
+                                            final Channel channel,
+                                            final SessionDisconnectHandler disconnectHandler,
+                                            final TransportType upgradedFromTransportType) throws Exception {
+    final String sessionId = connectPacket.getSessionId();
+    ManagedSession session = sessions.get(sessionId);
+    if (session == null) {
+      session = createSession(connectPacket, channel, disconnectHandler, upgradedFromTransportType);
+      ManagedSession fasterSession = sessions.putIfAbsent(sessionId, session);
+      if (fasterSession != null) {
+        session = fasterSession;
+      }
+    }
     return session;
   }
 
@@ -89,42 +99,27 @@ public class SessionStorage {
     final String origin = connectPacket.getOrigin();
     final String jsonpIndexParam = connectPacket.getJsonpIndexParam();
     final SocketAddress remoteAddress = connectPacket.getRemoteAddress();
-    try {
-      return sessionsMemoizer.get(sessionId,
-          new Computable<String, ManagedSession>() {
-            @Override
-            public ManagedSession compute(String sessionId) throws Exception {
-              if (transportType == TransportType.WEBSOCKET) {
-                return new WebSocketSession(channel, sessionId,
-                    origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
-              } else if (transportType == TransportType.FLASHSOCKET) {
-                return new FlashSocketSession(channel, sessionId,
-                    origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
-              } else if (transportType == TransportType.XHR_POLLING) {
-                return new XHRPollingSession(channel, sessionId,
-                    origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
-              } else if (transportType == TransportType.JSONP_POLLING) {
-                return new JsonpPollingSession(channel, sessionId,
-                    origin, disconnectHandler, upgradedFromTransportType, localPort, jsonpIndexParam, remoteAddress);
-              } else {
-                throw new UnsupportedTransportTypeException(transportType);
-              }
-            }
-          });
-    } catch (Exception e) {
-      throw new Exception(String.format(
-          "Failed to create new session: %s",
-          connectPacket.toString()), e);
+
+    // Create session by transport type
+    switch(transportType) {
+      case WEBSOCKET:
+        return new WebSocketSession(channel, sessionId,
+            origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
+      case FLASHSOCKET:
+        return new FlashSocketSession(channel, sessionId,
+            origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
+      case XHR_POLLING:
+        return new XHRPollingSession(channel, sessionId,
+            origin, disconnectHandler, upgradedFromTransportType, localPort, remoteAddress);
+      case JSONP_POLLING:
+        return new JsonpPollingSession(channel, sessionId,
+            origin, disconnectHandler, upgradedFromTransportType, localPort, jsonpIndexParam, remoteAddress);
+      default:
+        throw new UnsupportedTransportTypeException(transportType);
     }
   }
 
   public ManagedSession getSessionIfExist(final String sessionId) {
-    ManagedSession session = null;
-    try {
-      session = sessionsMemoizer.containsKey(sessionId) ? sessionsMemoizer
-          .get(sessionId) : null;
-    } catch (Exception ignore) {
-    }
-    return session;
+    return sessions.get(sessionId);
   }
 }